% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWBLRM-package.R
\docType{package}
\name{AWBLRM-package}
\alias{AWBLRM-package}
\alias{AWBLRM}
\title{The 'AWBLRM' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
