% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_dose.R
\name{decision.dose}
\alias{decision.dose}
\title{Next Recommended Dose}
\usage{
decision.dose(dose_levels, max_cycle, cycle_time, mode, target, od_control,
  cohort, dlt, dose, dlt_cycle, obs_time, mcmc_chains = 4,
  mcmc_samples = 3000)
}
\arguments{
\item{dose_levels}{A vector of length d, where d is the number of dose levels considered.}

\item{max_cycle}{Integer; Number of total treament cycles.}

\item{cycle_time}{Double; The length of a single treatment cycle.}

\item{mode}{Integer; Choose the decision rule to implement.}

\item{target}{The defined target probability of toxicity. For point-estimate-based decision rules, please input a single numeric value.
For interval-estimate-based decision rules, please input a vector indicating the target toxicity range, e.g. c(0.2, 0.33).}

\item{od_control}{Double(0-1); Escalation with overdose control.}

\item{cohort}{Vector; The cohort label of all patients currently observed.}

\item{dlt}{Vector; If toxic response was observed during the patient's evaluation window.}

\item{dose}{Vector;The dose level administered to the patients.}

\item{dlt_cycle}{Vector;Cycle in which DLT is observed for each patient. If no DLT please input -1.}

\item{obs_time}{Vector; How long has each patient been in the trial.}

\item{mcmc_chains}{Integer; Number of chains to run for MCMC sampling.}

\item{mcmc_samples}{Integer; Number of posterior samples to obtain}
}
\value{
A list with two elements. The first element is the next recommended dose. The second element is a summary information
matrix of the implemented decision rule.
}
\description{
Function to determine the next recommended dose according to the input data.
}
\examples{
#################################################################################
# Consider a 3-cycle treatment scenario with observed information of 9 patients #

all_doses = c(1, 2, 3, 4, 5)
max_cycle = 3
cycle_time = 30
mode = 1
target = 0.25
od_control = 1
cohort = c(1, 1, 1, 2, 2, 2, 3, 3, 3)
dlt = c(0, 0, 0, 1, 0, 0, 1, 0, 1)
dose = c(1, 1, 1, 2, 2, 2, 2, 2, 2)
dlt_cycle = c(-1, -1, -1, 2, -1, -1, 3, -1, 2)
obs_time = c(90, 45, 90, 35, 90, 75, 80, 90, 55)

test = decision.dose(all_doses, max_cycle, cycle_time, mode, target, od_control, cohort, dlt,
dose, dlt_cycle, obs_time)

}
