% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_dose.R
\name{run_stan2}
\alias{run_stan2}
\title{Bayesian Logistic Regression with Stan}
\usage{
run_stan2(mat, warmup, iter, chains, cores, cont)
}
\arguments{
\item{mat}{Current information matrix.}

\item{warmup}{Number of burn-in samples.}

\item{iter}{Number of posterior samples to obtain.}

\item{chains}{Number of chains to run.}

\item{cont}{Step size of the path search algorithm.}

\item{samp_cores}{Number of cores to perform sampling. Recommend to 1 if iter < 10000.}
}
\value{
The posterior samples of the slope and intercept.
}
\description{
Bayesian Logistic Regression with Stan
}
