% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{trial}
\alias{trial}
\title{Simulated Dose-finding Trial}
\usage{
trial(true, K, size, cycle_prob, cycle_time, target, samples, assessment,
  fixed = TRUE, mode = 1, stopping = 1, least = 12, dropout,
  od_control, mcmc_chains = 4, mcmc_samples = 3000)
}
\arguments{
\item{true}{Matrix; Pre-defined true dose-toxicity relationship. The input should be a 2-by-p matrix, with the 1st row being the
p dose levels and the 2nd row being the corresponding probability of DLT.}

\item{K}{Integer; Number of treatment cycles considered.}

\item{size}{Integer; Number of patients to enrolled in each cohort.}

\item{cycle_prob}{K-simplex; A vector that contains the true probability of DLT happening in each of the cycles. Sums up to one.}

\item{cycle_time}{Integer; Duration of a single treatment cycle (NOT entire treatment window!). Units are arbitrary.}

\item{target}{The defined target probability of toxicity. For point-estimate-based decision rules, please input a single numeric value.
For interval-estimate-based decision rules, please input a vector indicating the target toxicity range, e.g. c(0.2, 0.33).}

\item{samples}{Integer; The maximum number of patients to be enrolled into the trial. Also serves as the default stopping rule.}

\item{assessment}{Integer; The cycle in which the model is assessed to determine the dose level of the next cohort (1 means the model is assessed
after the completion of the 1st cycle of each cohort, 2 means 2nd cycle...).}

\item{fixed}{Boolean; TRUE if patients in the same cohort are enrolled at the same time. FALSE if the enrollment time of patients in the same
cohort follows a Uniform (0, cycle_time).}

\item{mode}{Integer; Determine which decision rule to implement to perform dose escalation / de-escalation. 1 = Original CRM, 2 = Inverse Point-Estimate,
3 = BLRM, 4 = Ratio. Defaults to 1. For details on the decision rule please refer to the report.}

\item{stopping}{Integer; Defined the stopping criteria. 1 = stop the trial when maximum enrollment is achieved, 2 = stop the trial if at least 12
patients are enrolled and 9 patients have already been treated at the next recommended dose. Defaults to 1.}

\item{least}{Integer; Least amount of patients to enrolled if using the second stopping rule. Defaults to 12, directly impacts "stopping".}

\item{dropout}{Boolean; TRUE if patient dropout time is assumened to follow an exponential distribution with a cumulative probability of 15% at the end
of the treatment window.}

\item{od_control}{Numeric (0 - 1); Escalation with overdose control. Escalate to the next dose if the probability of overdosing does not exceed the
input probability. No overdose control if od_control is set to 1.}

\item{mcmc_chains}{Integer; Number of chains to consider during MCMC sampling. Defaults to 3.}

\item{mcmc_samples}{Integer; Numbers of samples from the posterior distribution of the betas. Default to 3000 with a burn-in period of 500
samples.}
}
\value{
Returns a list with three information tables.

The first table gives the dose allocation of the patients during the trial(number of patients treated at each dose), including the corresponding
rate of DLT at each dose level.

The second table provides a summary table of the results after the trial is finished. This table includes the
number of patients treated, the duration of the trial, the number / proportion of DLTs, the proportion of patients allocated in dose levels
that are under / on / over the true MTD, and the final recommended dose.

The third table entails a full detailed information matrix of the trial.
}
\description{
This function simulates one dose-finding trial with the pre-specified design.
}
\examples{
# Set up the parameters for a simple simulation #
dose1 = rbind(1:5, c(0.05, 0.1, 0.25, 0.4, 0.6))
K1 = 2
size1 = 3
cycle_prob1 = c(0.5, 0.5)
cycle_time1 = 30
samples1 = 36
assessment1 = 1
fixed1 = TRUE
mode1 = 1
stopping1 = 1
least1 = 12
dropout1 = FALSE
od_control1 = 0.8
target1 = 0.25

# Simulate one trial according to the defined parameters. #
trial(dose1, K1, size1, cycle_prob1, cycle_time1, target1, samples1, assessment1,
fixed1, mode1, stopping1, least1, dropout1, od_control1)
}
